+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/vue
 
+## 4.0.97
+
+### Patch Changes
+
+- Updated dependencies [ef3bac4]
+- Updated dependencies [9942196]
+  - ai@7.0.97
+  - @ai-sdk/provider-utils@5.0.39
+
+## 4.0.96
+
+### Patch Changes
+
+- Updated dependencies [912fb01]
+- Updated dependencies [c595e6e]
+  - ai@7.0.96
+  - @ai-sdk/provider-utils@5.0.38
+
 ## 4.0.95
 
 ### Patch Changes